Moved the PromptsFromFileInvocation class from its own file (prompts_fromfile.py) )into the prompts.py file. Also cleaned up the code of the class a little bit.
